[all] Fix a bunch of warnings, try a different way of creating an allocator that maybe GCC will like

+++ include/nova_renderer/constants.hpp
@@ -16,7 +16,7 @@ namespace nova::renderer {
     constexpr uint32_t INTEL_PCI_VENDOR_ID = 8086;
     constexpr uint32_t NVIDIA_PCI_VENDOR_ID = 0x10DE;
 
-    constexpr uint8_t NUM_IN_FLIGHT_FRAMES = 3;
+    constexpr uint32_t NUM_IN_FLIGHT_FRAMES = 3;
 
     constexpr mem::Bytes PER_FRAME_MEMORY_SIZE = 2_mb;
